+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciEncryptionAlgorithmsTest.kt
@@ -4,6 +4,7 @@ import at.asitplus.KmmResult
 import at.asitplus.catching
 import at.asitplus.openid.RequestParameters
 import at.asitplus.openid.TokenResponseParameters
+import at.asitplus.signum.indispensable.josef.JweEncryption
 import at.asitplus.testballoon.withFixtureGenerator
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
 import at.asitplus.wallet.lib.agent.Holder
@@ -25,12 +26,15 @@ import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.assertions.throwables.shouldThrowAny
 import io.kotest.matchers.collections.shouldBeSingleton
 import io.kotest.matchers.nulls.shouldNotBeNull
+import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 
-val OidvciEncryptionTest by testSuite {
+val OidvciEncryptionAlgorithmsTest by testSuite {
 
     withFixtureGenerator {
         object {
+            val issuerEnc = JweEncryption.A128CBC_HS256
+            val walletEnc = JweEncryption.A128GCM
             val authorizationService = SimpleAuthorizationService(
                 strategy = CredentialAuthorizationServiceStrategy(setOf(ConstantIndex.AtomicAttribute2023)),
             )
@@ -43,7 +47,8 @@ val OidvciEncryptionTest by testSuite {
                 credentialSchemes = setOf(ConstantIndex.AtomicAttribute2023),
                 encryptionService = IssuerEncryptionService(
                     requireResponseEncryption = true, // this is important
-                    decryptionKeyMaterial = EphemeralKeyWithoutCert()
+                    decryptionKeyMaterial = EphemeralKeyWithoutCert(),
+                    supportedJweEncryptionAlgorithms = setOf(issuerEnc)
                 ),
             )
             val state = uuid4().toString()
@@ -51,6 +56,7 @@ val OidvciEncryptionTest by testSuite {
                 encryptionService = WalletEncryptionService(
                     requestResponseEncryption = true, // this is important
                     requireRequestEncryption = true, // this is important
+                    supportedJweEncryptionAlgorithm = walletEnc
                 )
             )
             val oauth2Client = OAuth2Client()
@@ -92,16 +98,19 @@ val OidvciEncryptionTest by testSuite {
                     credentialFormat = credentialFormat,
                     clientNonce = it.issuer.nonceWithDpopNonce().getOrThrow().response.clientNonce,
                 ).getOrThrow().shouldBeSingleton().first()
-                    .shouldBeInstanceOf<WalletService.CredentialRequest.Encrypted>(),
+                    .shouldBeInstanceOf<WalletService.CredentialRequest.Encrypted>().apply {
+                        request.header.encryption shouldBe it.issuerEnc
+                    },
                 credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
             ).getOrThrow().apply {
-                this.shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Encrypted>()
+                shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Encrypted>().apply {
+                    response.header.encryption shouldBe it.walletEnc
+                }
                 it.client.parseCredentialResponse(this, PLAIN_JWT, ConstantIndex.AtomicAttribute2023)
                     .getOrThrow().first().shouldBeInstanceOf<Holder.StoreCredentialInput.Vc>().apply {
                         signedVcJws.payload.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
                     }
             }
-
         }
 
         test("wallet does not encrypt credential request and decrypts credential response") {
@@ -123,92 +132,13 @@ val OidvciEncryptionTest by testSuite {
                     .shouldBeInstanceOf<WalletService.CredentialRequest.Plain>(),
                 credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
             ).getOrThrow().apply {
-                this.shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Encrypted>()
+                shouldBeInstanceOf<CredentialIssuer.CredentialResponse.Encrypted>().apply {
+                    response.header.encryption shouldBe it.walletEnc
+                }
                 it.client.parseCredentialResponse(this, PLAIN_JWT, ConstantIndex.AtomicAttribute2023)
                     .getOrThrow().first().shouldBeInstanceOf<Holder.StoreCredentialInput.Vc>().apply {
                         signedVcJws.payload.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
                     }
-            }
-
-        }
-
-        test("wallet does not encrypt credential request but issuer requires this") {
-            it.issuer = CredentialIssuer(
-                authorizationService = it.authorizationService,
-                issuer = IssuerAgent(
-                    identifier = "https://issuer.example.com".toUri(),
-                    randomSource = RandomSource.Default
-                ),
-                credentialSchemes = setOf(ConstantIndex.AtomicAttribute2023),
-                encryptionService = IssuerEncryptionService(
-                    requireResponseEncryption = true,
-                    decryptionKeyMaterial = EphemeralKeyWithoutCert(),
-                    requireRequestEncryption = true, // this is important for this test
-                ),
-            )
-
-            val requestOptions = WalletService.RequestOptions(ConstantIndex.AtomicAttribute2023, PLAIN_JWT)
-            val credentialFormat =
-                it.client.selectSupportedCredentialFormat(requestOptions, it.issuer.metadata)
-                    .shouldNotBeNull()
-            val scope = credentialFormat.scope.shouldNotBeNull()
-            val token = it.getToken(scope)
-
-            val request = it.client.createCredential(
-                tokenResponse = token,
-                // trick wallet into not encrypting
-                metadata = it.issuer.metadata.copy(credentialRequestEncryption = null),
-                credentialFormat = credentialFormat,
-                clientNonce = it.issuer.nonceWithDpopNonce().getOrThrow().response.clientNonce,
-            ).getOrThrow().shouldBeSingleton().first()
-                .shouldBeInstanceOf<WalletService.CredentialRequest.Plain>()
-
-            shouldThrow<OAuth2Exception.InvalidEncryptionParameters> {
-                it.issuer.credential(
-                    authorizationHeader = token.toHttpHeaderValue(),
-                    params = request,
-                    credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-                ).getOrThrow()
-            }
-
-        }
-
-        test("issuer fails to encrypt response") {
-            it.issuer = CredentialIssuer(
-                authorizationService = it.authorizationService,
-                issuer = IssuerAgent(
-                    identifier = "https://issuer.example.com".toUri(),
-                    randomSource = RandomSource.Default
-                ),
-                credentialSchemes = setOf(ConstantIndex.AtomicAttribute2023),
-                encryptionService = IssuerEncryptionService(
-                    requireResponseEncryption = true,
-                    encryptCredentialResponse = EncryptJweFun { header, payload, recipientKey ->
-                        KmmResult.catching { TODO("issuer fails to encrypt") }
-                    }
-                ),
-            )
-            val requestOptions = WalletService.RequestOptions(ConstantIndex.AtomicAttribute2023, PLAIN_JWT)
-            val credentialFormat =
-                it.client.selectSupportedCredentialFormat(requestOptions, it.issuer.metadata)
-                    .shouldNotBeNull()
-            val scope = credentialFormat.scope.shouldNotBeNull()
-            val token = it.getToken(scope)
-
-            val request = it.client.createCredential(
-                tokenResponse = token,
-                metadata = it.issuer.metadata,
-                credentialFormat = credentialFormat,
-                clientNonce = it.issuer.nonceWithDpopNonce().getOrThrow().response.clientNonce,
-            ).getOrThrow().shouldBeSingleton().first()
-                .shouldBeInstanceOf<WalletService.CredentialRequest.Encrypted>()
-
-            shouldThrowAny {
-                it.issuer.credential(
-                    authorizationHeader = token.toHttpHeaderValue(),
-                    params = request,
-                    credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-                ).getOrThrow()
             }
         }
     }